+++ app/src/main/AndroidManifest.xml
@@ -17,7 +17,7 @@
         android:theme="@style/Theme.Restaurants"
         tools:targetApi="31">
         <activity
-            android:name=".MainActivity"
+            android:name=".restaurants.presentation.MainActivity"
             android:exported="true"
             android:label="@string/app_name"
             android:theme="@style/Theme.Restaurants">
