+++ app/src/main/java/com/emma_ea/restaurants/restaurants/data/local/RestaurantsDao.kt
@@ -1,10 +1,12 @@
-package com.emma_ea.restaurants
+package com.emma_ea.restaurants.restaurants.data.local
 
 import androidx.room.Dao
 import androidx.room.Insert
 import androidx.room.OnConflictStrategy
 import androidx.room.Query
 import androidx.room.Update
+import com.emma_ea.restaurants.restaurants.data.local.LocalRestaurant
+import com.emma_ea.restaurants.restaurants.data.local.PartialLocalRestaurant
 
 @Dao
 interface RestaurantsDao {
