+++ app/src/main/java/com/emma_ea/restaurants/restaurants/domain/GetRestaurantByIDUseCase.kt
@@ -1,0 +1,13 @@
+package com.emma_ea.restaurants.restaurants.domain
+
+import com.emma_ea.restaurants.restaurants.data.RestaurantRepository
+
+class GetRestaurantByIDUseCase {
+
+    private val repository = RestaurantRepository()
+
+    suspend operator fun invoke(id: Int): Restaurant {
+        return repository.getRestaurantById(id)
+    }
+
+}
