+++ app/src/main/java/com/emma_ea/restaurants/restaurants/domain/Restaurant.kt
@@ -1,4 +1,4 @@
-package com.emma_ea.restaurants
+package com.emma_ea.restaurants.restaurants.domain
 
 data class Restaurant(
     val id: Int,
