+++ app/src/main/java/com/emma_ea/restaurants/restaurants/domain/ToggleRestaurantUseCase.kt
@@ -1,4 +1,6 @@
-package com.emma_ea.restaurants
+package com.emma_ea.restaurants.restaurants.domain
+
+import com.emma_ea.restaurants.restaurants.data.RestaurantRepository
 
 class ToggleRestaurantUseCase {
