+++ app/src/main/java/com/emma_ea/restaurants/restaurants/presentation/list/RestaurantViewState.kt
@@ -1,4 +1,6 @@
-package com.emma_ea.restaurants
+package com.emma_ea.restaurants.restaurants.presentation.list
+
+import com.emma_ea.restaurants.restaurants.domain.Restaurant
 
 data class RestaurantViewState(
     val data: List<Restaurant> = emptyList(),
